requirements.devel.txt
Flask==1.1.2
pytest==5.4.1
pytest-env==0.6.2
pytest-html==2.1.1
pytest-metadata==1.8.0
flake8==3.7.9
flake8-html==0.4.0
requests-mock
pytest-asyncio
pytest-aiohttp